requirements.txt
# MCP & server
fastmcp>=2.14.0,<3
ollama>=0.4.0

# Vector DB & embeddings
chromadb>=0.5.0

# PDF ingestion & generation
pypdf>=5.0.0
fpdf2>=2.8.0

# Config & validation
python-dotenv>=1.0.0
pydantic>=2.0.0
